Add a queue to download HTML content.
This commit also adds dependencies to pika (for rabbitmq management) and to chardet to autodetect the charset of some pages.

If some resources aren't detectable, they are blacklisted and will not be considered in the recommendation process nor tried again in the future.

This commit also adds a progress bar to be able to know what's happening when processing the information.