Hurray, a recommander which recommands!
A lot of things happening with this commit. Mainly, that's because I was too lazy to commit regularly. As I'm the only one working on this project, that's not really a problem, thus.

The recommandation engine is now able to provide recommendations. The recommendations are for now only based on the TF-IDF scores. Work is on its way to support different processors (context, using profiles, etc).

This commit brings with it other changes, such as:

* a complete refactoring of profiles.py, which is actually not needed anymore. The code is now embedded into classes and methods, so it's way easier to use it.
* All the high level code is bundled together at run.py, in a class named Infuse. This class has several methods, for the different things you may want to do with it.

Thus, there is still a lot of work to do. The recommendations I currenly have are not so good, and this for multiple reasons:

1. they are too numerous. I need a way to filter them, as of now, it's just a list of ranked resources. As I didn't finished the "feedback" website yet, people (appart from me) didn't provided me feedback. This need to change.
2. I need to test another similarity measures: euclidean distance is good, but there should be better options
3. should use different set of features (currently I'm not using something else that text feature for the POC)
4. rankings could be computed in a different way, using assumptions

So, lot more stuff in the pipes, but I got something working, YAY !